Issue with UI thread switching fixed
Fixed an issue where the app would crash when you switch between
landscape and portrait orientation. This is due to the fact that the
list of test results is updated on a background thread and the
listadapter is updated on the ui thread. both need to be updated on the
UI thread, so I moved a bit of code around to fix the issue